refactor: split api/services/settings.py by domain (CHAOS-1544)
Split the 1622-line api/services/settings module into a focused package
with one submodule per service domain. All existing public symbols are
re-exported from settings/__init__.py so every consumer keeps working
without changes.

New layout:
- settings/_helpers.py            credential-key normalization + lazy
                                  admin.schemas imports
- settings/generic.py             SettingsService (key/value)
- settings/integration_credentials.py  IntegrationCredentialsService
- settings/sync_configuration.py  SyncConfigurationService
- settings/identity_mapping.py    IdentityMappingService
- settings/team_mapping.py        TeamMappingService
- settings/team_discovery.py      TeamDiscoveryService
- settings/team_drift_sync.py     TeamDriftSyncService
- settings/team_membership.py     TeamMembershipService
- settings/jira_activity_inference.py  JiraActivityInferenceService

Pure refactor: every method body is preserved verbatim. The encryption
helpers (decrypt_value/encrypt_value) and _normalize_credential_keys
are still importable from dev_health_ops.api.services.settings so SSO
and credential-resolver imports continue to resolve.

The package name 'settings' is misleading (this is not Pydantic
BaseSettings) but renaming is high-risk because of mock.patch paths in
~25 test files. Tracked as CHAOS-1559 for follow-up.
